1. Убран отдельный blueprint под страницу правления, перемещен в папку ЛК. 2. HTML-страницы разбиты по папкам. 3. Создан HTML-файл menu. 4. Переделаны HTML-страницы: убран повторяющийся код для меню. 5. Переделана логика кнопок меню: войти/зарегистрироваться; личный кабинет/выйти. 6. Для админа предусмотрена возможность перемещаться по ЛК владельцев участков. 7. Переделан логин для админа и пользователей. Теперь пользователь не сможет попасть на чужую страницу ЛК. Админ может попасть на любую страницу. П.С.: переделать вывод ошибок-flash не представляется возможным. Ломается вся страница регистрации (связано с тем, что это единый объект bootstrap). Что еще можно улучшить в проекте?
