+++ webapp/templates/login/registration.html
@@ -1,7 +1,9 @@
 {% extends "base_login.html" %}
 
 {% block body %}
+
 <body class="d-flex align-items-center py-4 bg-body-tertiary">
+
 <main class="form-signin w-100 m-auto">
      {% with messages = get_flashed_messages() %}
         {% if messages %}
